finalising the ridewise application

+++ ride-wise-cab-sharing/src/main/java/com/example/RideWise/ride/wise/cab/sharing/Service/RiderService.java
@@ -5,6 +5,7 @@ import com.example.RideWise.ride.wise.cab.sharing.Dto.RiderDetailsDto;
 import com.example.RideWise.ride.wise.cab.sharing.Entity.Ride;
 import com.example.RideWise.ride.wise.cab.sharing.Entity.Rider;
 import com.example.RideWise.ride.wise.cab.sharing.Entity.User;
+import com.example.RideWise.ride.wise.cab.sharing.Enum.Role;
 import com.example.RideWise.ride.wise.cab.sharing.Exceptions.DriverNotFoundException;
 import com.example.RideWise.ride.wise.cab.sharing.Exceptions.RiderAlreadyExistsException;
 import com.example.RideWise.ride.wise.cab.sharing.Exceptions.RiderNotFoundException;
@@ -12,6 +13,7 @@ import com.example.RideWise.ride.wise.cab.sharing.Repository.RideRepository;
 import com.example.RideWise.ride.wise.cab.sharing.Repository.RiderRepository;
 import com.example.RideWise.ride.wise.cab.sharing.Repository.UserRepository;
 import jakarta.transaction.Transactional;
+import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.Nullable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +24,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
 public class RiderService {
 
     @Autowired
@@ -30,15 +33,15 @@ public class RiderService {
     @Autowired
     private UserRepository userRepository;
 
-    @Autowired
-    private RideRepository rideRepository;
+   // @Autowired
+   // private RideRepository rideRepository;
 
     public List<Rider> getAllRiders() {
         return riderRepository.findAll();
     }
 
     public RiderDetailsDto getRiderInfo(User customUser) throws RiderNotFoundException {
-        Rider rider = riderRepository.findByUser(customUser);
+        Rider rider = riderRepository.findByUser(customUser).orElseThrow(() -> new RiderNotFoundException("rider with email " + customUser.getEmail() + " not found"));
         if (rider.getRides().isEmpty()) {
             return RiderDetailsDto.builder()
                     .id(rider.getId())
@@ -46,6 +49,7 @@ public class RiderService {
                     .lastName(rider.getLastName())
                     .email(customUser.getEmail())
                     .completedRides(null)
+                    .memberSince(rider.getMemberSince().getYear())
                     .build();
         }
         return RiderDetailsDto.builder()
@@ -54,25 +58,22 @@ public class RiderService {
                 .lastName(rider.getLastName())
                 .email(customUser.getEmail())
                 .completedRides(rider.getRides())
+                .memberSince(rider.getMemberSince().getYear())
                 .build();
     }
 
-//    @Transactional
-//    public Rider createNewRider(Rider rider) throws RiderAlreadyExistsException {
-//        if (!riderRepository.existsByEmail(rider.getEmail())) {
-//            riderRepository.save(rider);
-//        }
-//        throw new RiderAlreadyExistsException("Rider already exists!");
-//    }
-
     @Transactional
     public String deleteRider(User customUser) throws RiderNotFoundException {
-        try {
-            Rider rider = userRepository.findByUser(customUser);
+        Rider rider = userRepository.findByUser(customUser).orElseThrow(() -> new RiderNotFoundException("rider with email " + customUser.getEmail() + " not found"));
+        User user = rider.getUser();
+        if (user.getRole().contains(Role.DRIVER)) {
+            user.getRole().remove(Role.RIDER);
+            userRepository.save(user);
             riderRepository.delete(rider);
-            return "Rider deleted successfully";
-        } catch (Exception e) {
-            throw new RiderNotFoundException("Driver not found!");
+            return "Rider deleted successfully!";
         }
+        userRepository.delete(user);
+        riderRepository.delete(rider);
+        return "Rider deleted successfully!";
     }
 }
